Change characters in Group type from Set to List
Unfortunately, when Spring MongoDb persists an object, the result
is a clone of the original one, giving different hashCodes.

This gives problems when attempting to see if a preexisting character
 is in the Set, as the original hashCode won't match the cloned one,
 the result will always be false.

Fix to this, is change type to List -> ArrayList, which don't have this
problem.

But having lost the uniqueness property of a Set, a CharacterAlreadyInGroupException
has been added to consider the repeated characters case.


Former-commit-id: be36a4bb0fc52a55ab734fb176bfa722b42f8536